Add files via upload
An Age Identifier System: Built with HTML, CSS, and JavaScript is a web-based tool that calculates a user's age based on their inputted birthdate. The HTML provides the structure, including a date input field and a button, while CSS enhances the visual design with styling for a responsive and attractive interface. While JavaScript handles the core logic.